requirements.txt
python-dotenv
langchain
langchain_community
langchain_core
langchain_text_splitters
langserve
fastapi
uvicorn
sse_starlette
pymongo
fastapi-jwt-auth
pydantic



